+++ CHANGELOG.md
@@ -140,3 +140,7 @@
 * Added option for ApplyThemeScript to apply when theme policy is ExplicitOnly or not
 * Cleaned up handling of base frame in dialogue scene (now available via DialogueScene.BaseFrame)
 * Suppressed warnings from ExplicitKBMInputMapper when noclipping (and other "missing axis" warnings as well)
+* Changed ammo type backing type from AmmoType enum to string
+* Added lock time to weapons [WIP]
+* Added burst fire to weapons [WIP]
+* Extended projectile and explosion handling for weapons [WIP]